Chore: Move SubjectTuple and SubjectType to Subjects definition file
