+++ src/base/v20170710.ts
@@ -49,33 +49,6 @@ export type Level = number & {};
 export const Level = v.pipe(v.number(), v.minValue(MIN_LEVEL), v.maxValue(MAX_LEVEL));
 
 /**
- * The types of subjects used on WaniKani and its API.
- *
- * @see {@link https://docs.api.wanikani.com/20170710/#subjects}
- * @category Base
- * @category Subjects
- */
-export type SubjectType = "kana_vocabulary" | "kanji" | "radical" | "vocabulary";
-export const SubjectType = v.picklist(["kana_vocabulary", "kanji", "radical", "vocabulary"]);
-
-/**
- * A non-empty array of WaniKani subject types.
- *
- * @see {@link https://docs.api.wanikani.com/20170710/#subjects}
- * @category Base
- * @category Subjects
- */
-export type SubjectTuple = [SubjectType, ...SubjectType[]];
-export const SubjectTuple = v.pipe(
-  v.tupleWithRest([SubjectType], SubjectType),
-  v.nonEmpty(),
-  v.checkItems(
-    (item, index, array) => array.indexOf(item) === index,
-    "Duplicate Subject Type detected in Subject Tuple",
-  ),
-);
-
-/**
  * The common properties across all Resources from the WaniKani API.
  *
  * @remarks This is a partial interface; most use cases involve using a reource that extends it.
