+++ tests/base/v20170710.test-d.ts
@@ -4,8 +4,6 @@ import {
   type CollectionParameters,
   type DatableString,
   type Level,
-  type SubjectTuple,
-  type SubjectType,
   stringifyParameters,
 } from "../../src/base/v20170710.js";
 import { assertType, describe, expectTypeOf } from "vitest";
@@ -32,28 +30,6 @@ describe("Level", () => {
     } else {
       throw new TypeError("Expected levels to be an array");
     }
-  });
-});
-
-describe("SubjectType", () => {
-  testFor("Valid Subject Types", ({ subjectTypes }) => {
-    if (Array.isArray(subjectTypes)) {
-      subjectTypes.forEach((subject) => {
-        assertType<SubjectType>(subject);
-      });
-    } else {
-      throw new TypeError("Expected subjectTypes to be an array");
-    }
-  });
-});
-
-describe("SubjectTuple", () => {
-  // These tests are kinda redundant, but we'll leave them here for completeness' sake
-  testFor("Partial SubjectTuple is Valid", ({ partialSubjectTuple }) => {
-    assertType<SubjectTuple>(partialSubjectTuple);
-  });
-  testFor("Full SubjectTuple is valid", ({ fullSubjectTuple }) => {
-    assertType<SubjectTuple>(fullSubjectTuple);
   });
 });
